1_company-list/requirements.txt
crewai
langchain-google-genai
python-dotenv
pydantic>=2.0.0
